rtc: rv3028: Write BSM and TCE/TCR to EEPROM
Periodically the RV3028 refreshes registers from the EEPROM. When this
happens, some settings that have only been committed to registers are
lost. Change the handling of backup-switchover-mode and
trickle-resistor-ohms to write the EEPROM instead (if something has
changed), on the understanding that registers will be refreshed
afterwards.

See: https://github.com/raspberrypi/linux/issues/2912

Signed-off-by: Phil Elwell <phil@raspberrypi.com>