DTS: bcm2712: don't assume L1 sub-state support at chip level
L1 sub-states require the connection of the CLKREQ# auxiliary signal
which is a PCB-level decision, and also depends on the vagrancies of
adapters/interposers/slots that may exist between the chip and the
endpoint.

If this parameter is present, and the RC driver sees an L1 substate
capability in the endpoint, then the result is a broken link if CLKREQ#
isn't end-to-end connected - regardless of the state of the L1.x enable
bits in the RC.

Board-specific methods (HAT+ overlay, board DTB) should define this
property if required.

Signed-off-by: Jonathan Bell <jonathan@raspberrypi.com>