ASoC: bcm2835-i2s: Set the PERIOD_BYTES min to 256
Commit [1] set the minimum PERIOD_BYTES value to the product of the DMA
burst size and 8. For the I2S interface on BCM2835 that equates to 16
where it used to be 256. ffmpeg uses the minimum value as its preferred
value, leading to many, many very small periods, which affects
performance and leads to complaints about DTS timestamps.

Restore the previous behaviour and performance by making the bcm2835-i2s
driver set a minimum PERIOD_BYTES value of 256.

Link: https://github.com/raspberrypi/linux/issues/5709

[1] 300689fb04b3 ("ASoC: soc-generic-dmaengine-pcm: set
period_bytes_min based on maxburst")

Signed-off-by: Phil Elwell <phil@raspberrypi.com>